+++ playground/src/main.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 import { createRoot } from 'react-dom/client';
 
 import { App } from './App';
-import './styles.css';
+import './index.css';
 
 createRoot(document.getElementById('root')!).render(
   <React.StrictMode>